requirements.txt
flask>=2.0.1
requests>=2.26.0
twilio>=7.0.0
python-dotenv==1.0.0
ngrok==0.12.0 